[gitpod-network-check] Refactor to extract tests and runner
Tool: gitpod/catfood.gitpod.cloud
